# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simple Shop API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List of products",
                "description": "Get all products",
                "operationId": "getProducts",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Potato"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
                                        },
                                        "price": {
                                            "type": "float",
                                            "example": "123.45"
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PLN"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{sort}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List of products",
                "description": "Get list of products sorted by a column of your choice (ID is the default)",
                "operationId": "getProductsSorted",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "path",
                        "description": "Sort by a column of your choice (id, name, description, price or currency)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Potato"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
                                        },
                                        "price": {
                                            "type": "float",
                                            "example": "123.45"
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PLN"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/show/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Details of a product",
                "description": "Get one of products",
                "operationId": "getProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of a product",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Potato"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
                                        },
                                        "price": {
                                            "type": "float",
                                            "example": "123.45"
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PLN"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/create": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a Product",
                "description": "Create a product by providing a name, description, price and currency",
                "operationId": "createProduct",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Potato"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
                                    },
                                    "price": {
                                        "type": "float",
                                        "example": "123.45"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "PLN"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Added Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Result": {
                                            "type": "string",
                                            "example": "Product has been Added"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/update/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "description": "Update one of products",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of a product",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Potato"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam"
                                    },
                                    "price": {
                                        "type": "float",
                                        "example": "123.45"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "PLN"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product Added Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Result": {
                                            "type": "string",
                                            "example": "Product has been updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/delete/{id}": {
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "description": "Delete one of products",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of a product",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product Added Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Result": {
                                            "type": "string",
                                            "example": "Product has been deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}